# Optimize `sendfile`