## Description of the changes <!-- (reasons and measures) -->

I looked into why the `sendfile` LTP test is so slow, and found two things:
* the buffer size is pretty small,
* `DkEventSet` in Linux PAL is pretty inefficient and always calls `futex` (Linux-SGX PAL has a waiters count, so it should be OK)

As a result, there is a lot of small `pread64` and `pwrite64` calls, each accompanied by several `futex` calls (releasing locks).

Here's a benchmark for `sendfile09` LTP test (which uses `sendfile` to copy 1 GB ranges), with `gramine-direct`, compiled in debug mode (but release numbers are pretty similar):

| version | time |
|---|---|
| Linux (no Gramine) | 1.07 s |
| master |  2.61 s |
| `DkEventSet` optimized | 1.86 s |
| 64KB buffer | 1.31 s |
| **`DkEventSet` optimized + 64KB buffer** | **1.25 s** |

(Benchmarks performed using [hyperfine](https://github.com/sharkdp/hyperfine), it's pretty neat).

## How to test this PR? <!-- (if applicable) -->

Please read the DkEventSet code carefully.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/gramineproject/gramine/245)
<!-- Reviewable:end -->